+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add `Element#addIntervalListener()` to add an interval listener that will only run while the element is connected
 * Ignore history change when `renderer.history` is explicitly false
 * Fix `Scene#enableStyle()` creating too many pledges
+* Make `break number` actually work as expected
 
 ## 2.2.16 (2022-10-08)